reserve and resize of the property vectors now take size_t
git-svn-id: http://openmesh.org/svnrepo/OpenMesh@840 fdac6126-5c0c-442c-9429-916003d36597

+++ trunk/src/OpenMesh/Core/Mesh/BaseKernel.hh
@@ -638,8 +638,8 @@ public: //----------------------------------------------------- element numbers
 
 protected: //------------------------------------------- synchronize properties
 
-  void vprops_reserve(unsigned int _n) const { vprops_.reserve(_n); }
-  void vprops_resize(unsigned int _n) const { vprops_.resize(_n); }
+  void vprops_reserve(size_t _n) const { vprops_.reserve(_n); }
+  void vprops_resize(size_t _n) const { vprops_.resize(_n); }
   void vprops_clear() {
     vprops_.clear();
   }
@@ -647,8 +647,8 @@ protected: //------------------------------------------- synchronize properties
     vprops_.swap(_i0, _i1);
   }
 
-  void hprops_reserve(unsigned int _n) const { hprops_.reserve(_n); }
-  void hprops_resize(unsigned int _n) const { hprops_.resize(_n); }
+  void hprops_reserve(size_t _n) const { hprops_.reserve(_n); }
+  void hprops_resize(size_t _n) const { hprops_.resize(_n); }
   void hprops_clear() {
     hprops_.clear();
   }
@@ -656,8 +656,8 @@ protected: //------------------------------------------- synchronize properties
     hprops_.swap(_i0, _i1);
   }
 
-  void eprops_reserve(unsigned int _n) const { eprops_.reserve(_n); }
-  void eprops_resize(unsigned int _n) const { eprops_.resize(_n); }
+  void eprops_reserve(size_t _n) const { eprops_.reserve(_n); }
+  void eprops_resize(size_t _n) const { eprops_.resize(_n); }
   void eprops_clear() {
     eprops_.clear();
   }
@@ -665,8 +665,8 @@ protected: //------------------------------------------- synchronize properties
     eprops_.swap(_i0, _i1);
   }
 
-  void fprops_reserve(unsigned int _n) const { fprops_.reserve(_n); }
-  void fprops_resize(unsigned int _n) const { fprops_.resize(_n); }
+  void fprops_reserve(size_t _n) const { fprops_.reserve(_n); }
+  void fprops_resize(size_t _n) const { fprops_.resize(_n); }
   void fprops_clear() {
     fprops_.clear();
   }
@@ -674,7 +674,7 @@ protected: //------------------------------------------- synchronize properties
     fprops_.swap(_i0, _i1);
   }
 
-  void mprops_resize(unsigned int _n) const { mprops_.resize(_n); }
+  void mprops_resize(size_t _n) const { mprops_.resize(_n); }
   void mprops_clear() {
     mprops_.clear();
   }
